// src/app/models/add-products.model.ts
export interface Products {

    id: number;
    uniqueId: string;
    sku: string;
    name: string;
    description: string;
    imageUrl: string;
    contentUrl: string;
    price: number;
    status: string;
    enabled: string;
    openPrice: boolean;
    openStock: boolean;
    payload:string;
    qcode: string;   
     
}